test shell quoting separately from normal local::lib shell directives
Some shells have special handling for some env vars, such as rejecting
nonexistent directories in PATH.  This is an issue for the fish shell.
We want to test the quoting routines even on strings that aren't valid
paths, so separate out the testing of odd strings from the testing for
local::lib's normal env variables.