Back-compatability workaround for CPAN.pm bootstrap
Older CPAN::FirstTime, e.g. as ship with 5.18, don't work with
the current local::lib because of the internal changes around
activation.

This patch zaps PERL_LOCAL_LIB_ROOT in environment_vars_string_for() only
if the caller is an older CPAN::FirstTime.  This restores the ability of
CPAN::FirstTime to bootstrap local::lib and shouldn't affect any other
uses of local::lib.

I have tested this manually on both darwin and linux on a stock
5.18.