create directory structure under local::lib including arch dir
perl will include the arch dir under a PERL5LIB path in @INC
automatically, but only if it exists.  For programs that install
modules, the arch path might not exist when started.

When initializing a local::lib, create the arch (and other)
subdirectories to make sure they always end up in @INC.