fix fish environment escaping
fish 2.2 changed how it handles array variables without a mention in its
changelog.

For array environment variables, they are stringified separated with
spaces within fish itself.  The value stored in the environment is
joined with no spaces.  Since we can't control this, we can't use array
variables for most environment variables.

PATH, CDPATH, and MANPATH work differently.  The value stored in the
environment is joined with colons.  fish expects these variables to be
arrays for internal functions, so we can't just handle them as strings.
Also, PATH and CDPATH are validated against the filesystem when being
set, so we need to silence any errors from missing directories.

This requires two separate quoting mechanisms.  One specifically for
PATH, CDPATH, and MANPATH, handling them as arrays.  And another for
everything else, handling them as strings.