only use strict and warnings under RELEASE_TESTING
This generally makes us leaner, but also on perl 5.8, warnings loads
Carp (and thus Exporter).  This is a problem because users can have
updated versions of the modules in their local::lib.  In addition, Carp
used to be split into Carp and Carp::Heavy.  A new Carp will have an
empty Carp::Heavy, thus leaving it in a broken state if the two are
mismatched.