better account for powershell oddities
powershell 1 doesn't support -ExecutionPolicy, or have a per-process
mechanism for setting the execution process.  Just skip running on it.

The execution policy can also be overridden by a system or domain admin,
so we need to make sure we can set it before trying to use powershell at
all.

Also need some extra bits to actually get error codes out of the
process.