avoid spaces in temp directory names
If there is a space in the created temp directory, and it is mishandled
by a program somewhere, it will often result in files being created
outside the temp directory.  The files will then not get properly
cleaned up.

Instead, create a temp directory without spaces, and create the
directory with spaces inside that.