remove env vars rather leaving undef to avoid taint
localizing %ENV entries on its own would be setting them to undef. But
since they are environment variables, undef would be converted to an
empty string. An empty string in PATH is the same as the current
directory. This should trip taint checks, and will in the future.

Instead of leaving as undef, delete the %ENV entries after localizing
them. When deleted, they do not act as empty strings.